fix bug that ignored ingresses on requests
The bug in question could be triggered by attempting to request a
distribution that was built with webpacks publicPath set to one of
the ingresses. The solution seemed to be removing ingresses' paths
from req.baseUrl prior to fetching the files. Though this patches
the problem, a better solution should be applied in the future.